Move unweildy match logic into new mod
