+++ zingolib/src/wallet.rs
@@ -69,6 +69,7 @@ pub mod keys;
 #[cfg(not(feature = "integration_test"))]
 pub(crate) mod keys;
 pub(crate) mod message;
+mod send;
 pub(crate) mod traits;
 pub(crate) mod transactions;
 pub(crate) mod utils;
@@ -1155,11 +1156,7 @@ impl LightWallet {
     async fn select_notes_and_utxos(
         &self,
         policy: NoteSelectionPolicy,
-        PoolTargets {
-            target_transparent,
-            target_sapling,
-            target_orchard,
-        }: PoolTargets,
+        pool_targets: PoolTargets,
     ) -> Result<
         (
             Vec<SpendableOrchardNote>,
@@ -1169,74 +1166,18 @@ impl LightWallet {
         ),
         NoteSelectionError,
     > {
-        if target_transparent != Amount::zero()
+        if pool_targets.target_transparent != Amount::zero()
             && policy != NoteSelectionPolicy::AllowRevealedRecipients
             && !(policy < NoteSelectionPolicy::AllowRevealedRecipients)
         {
             return Err(NoteSelectionError::DisallowedByPrivacyPolicy);
         }
         match policy {
-            NoteSelectionPolicy::FullPrivacy => {
-                let sapling_candidates = self
-                    .get_spendable_domain_specific_notes::<SaplingDomain<zingoconfig::ChainType>>()
-                    .await;
-                let (sapling_notes, sapling_value_selected) = Self::add_notes_to_total::<
-                    SaplingDomain<zingoconfig::ChainType>,
-                >(
-                    sapling_candidates, target_sapling
-                );
-                let orchard_candidates = self
-                    .get_spendable_domain_specific_notes::<OrchardDomain>()
-                    .await;
-                let (orchard_notes, orchard_value_selected) =
-                    Self::add_notes_to_total::<OrchardDomain>(orchard_candidates, target_orchard);
-
-                if orchard_value_selected >= target_orchard
-                    && sapling_value_selected >= target_sapling
-                {
-                    Ok((
-                        orchard_notes,
-                        sapling_notes,
-                        Vec::new(), // No utxos for full privacy
-                        (sapling_value_selected + orchard_value_selected).unwrap(),
-                    ))
-                } else {
-                    Err(NoteSelectionError::InsufficientPrivateFunds)
-                }
-            }
+            NoteSelectionPolicy::FullPrivacy => self.select_notes_full_privacy(pool_targets).await,
             NoteSelectionPolicy::AllowRevealedAmounts => {
-                let full_target = (target_sapling + target_orchard).unwrap();
-                let sapling_candidates = self
-                    .get_spendable_domain_specific_notes::<SaplingDomain<zingoconfig::ChainType>>()
-                    .await;
-                let (sapling_notes, sapling_value_selected) = Self::add_notes_to_total::<
-                    SaplingDomain<zingoconfig::ChainType>,
-                >(
-                    sapling_candidates, full_target
-                );
-                if sapling_value_selected >= full_target {
-                    return Ok((
-                        Vec::new(),
-                        sapling_notes,
-                        Vec::new(),
-                        sapling_value_selected,
-                    ));
-                }
-                let orchard_candidates = self
-                    .get_spendable_domain_specific_notes::<OrchardDomain>()
-                    .await;
-                let (orchard_notes, orchard_value_selected) =
-                    Self::add_notes_to_total::<OrchardDomain>(
-                        orchard_candidates,
-                        (full_target - sapling_value_selected).unwrap(),
-                    );
-                Ok((
-                    orchard_notes,
-                    sapling_notes,
-                    Vec::new(),
-                    (sapling_value_selected + orchard_value_selected).unwrap(),
-                ))
+                self.select_notes_revealed_amounts(pool_targets).await
             }
+
             _ => todo!(),
         }
     }
